+++ website/docusaurus.config.js
@@ -37,6 +37,20 @@ const config = {
       }),
     ],
   ],
+  themes: [
+    [
+      require.resolve('@easyops-cn/docusaurus-search-local'),
+      /** @type {import("@easyops-cn/docusaurus-search-local").PluginOptions} */
+      ({
+        hashed: true,
+        language: ['en', 'zh'],
+        docsDir: '../docs',
+        docsRouteBasePath: 'docs',
+        indexBlog: false,
+        highlightSearchTermsOnTargetPage: true,
+      }),
+    ],
+  ],
   themeConfig:
     /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
     ({